Preparing for web deployment
This commit adds the necessary files for deploying the sudoku solver to the web. The cpp app was compiled to wasm using emscripten. The js file is used to load the wasm file and run the solver. The html file is used to display the sudoku board and the solver. The serve.py file is used to serve the files locally for testing.